requirements.txt
aiogram==3.4.1
torch
torchvision
timm
openai
Pillow
httpx
python-dotenv
